接口完善

+++ src/com/my/spring/serviceImpl/MessageFileServiceImpl.java
@@ -1,12 +1,20 @@
 package com.my.spring.serviceImpl;
 
+import com.my.spring.DAO.FileDao;
 import com.my.spring.DAO.MessageFileDao;
 import com.my.spring.enums.ErrorCodeEnum;
+import com.my.spring.enums.UserTypeEnum;
+import com.my.spring.model.Files;
 import com.my.spring.model.MessageFile;
+import com.my.spring.model.User;
+import com.my.spring.service.FileService;
 import com.my.spring.service.MessageFileService;
 import com.my.spring.utils.DataWrapper;
+import com.my.spring.utils.SessionManager;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -17,28 +25,81 @@ import java.util.List;
 public class MessageFileServiceImpl implements MessageFileService {
     @Autowired
     MessageFileDao messageFileDao;
+    @Autowired
+    FileDao fileDao;
+    @Autowired
+    FileService fileService;
+    private String filePath="D:\\fileupload";
+    private Integer fileType=3;
     @Override
-    public DataWrapper<Void> addMessageFile(MessageFile MessageFile,String token) {
-        DataWrapper<Void> dataWrapper = new DataWrapper<Void>();
-        if(!messageFileDao.addMessageFile(MessageFile)) {
-            dataWrapper.setErrorCode(ErrorCodeEnum.Error);
-        }
+    public DataWrapper<Void> addMessageFile(MessageFile messageFile,String token,MultipartFile file) {
+    	DataWrapper<Void> dataWrapper = new DataWrapper<Void>();
+        User userInMemory = SessionManager.getSession(token);////验证登录时的session
+        if (userInMemory != null) {
+			if(userInMemory.getUserType()==UserTypeEnum.Admin.getType()){
+				///////验证是不是管理员身份
+				if(messageFile!=null){////验证上传的实体类是不是为空
+					///////1.文件的上传返回url
+					String path=filePath+"/"+"messages"+"/";
+					String newfilename=fileService.uploadFile(path, file,fileType);
+					DataWrapper<Files> dataWrappers=fileDao.getByName(newfilename);
+					messageFile.setFileId(dataWrappers.getData().getId());
+					if(!messageFileDao.addMessageFile(messageFile)) 
+			            dataWrapper.setErrorCode(ErrorCodeEnum.Error);
+					else
+						return dataWrapper;
+			        
+				}else{
+					dataWrapper.setErrorCode(ErrorCodeEnum.Empty_Inputs);
+				}
+			}else{
+				dataWrapper.setErrorCode(ErrorCodeEnum.AUTH_Error);
+			}
+		} else {
+			dataWrapper.setErrorCode(ErrorCodeEnum.User_Not_Logined);
+		}
         return dataWrapper;
     }
 
     @Override
-    public DataWrapper<Void> deleteMessageFile(Long id,String token) {
+    public DataWrapper<Void> deleteMessageFile(Long id,Long fileid,String token) {
         DataWrapper<Void> dataWrapper = new DataWrapper<Void>();
-        if(!messageFileDao.deleteMessageFile(id)) {
-            dataWrapper.setErrorCode(ErrorCodeEnum.Error);
+        User userInMemory=SessionManager.getSession(token);
+        if(userInMemory!=null){
+        	if(userInMemory.getUserType()==UserTypeEnum.Admin.getType()){
+				///////验证是不是管理员身份
+				if(id!=null){////验证上传的实体类是不是为空
+					if(!messageFileDao.deleteMessageFile(id)) {
+						fileDao.deleteFiles(fileid);//删除文件表的信息
+						Files files=fileDao.getById(id);
+						fileService.deleteFileByPath(files.getUrl());///删除实际文件
+			            dataWrapper.setErrorCode(ErrorCodeEnum.Error);
+			        }
+				}else{
+					dataWrapper.setErrorCode(ErrorCodeEnum.Empty_Inputs);
+				}
+			}else{
+				dataWrapper.setErrorCode(ErrorCodeEnum.AUTH_Error);
+			}
+        }else{
+        	dataWrapper.setErrorCode(ErrorCodeEnum.User_Not_Logined);
         }
+        
         return dataWrapper;
     }
 
 
     @Override
     public DataWrapper<List<MessageFile>> getMessageFileList(String token ) {
-        return messageFileDao.getMessageFileList();
+    	DataWrapper<List<MessageFile>> dataWrapper = new DataWrapper<List<MessageFile>>();
+    	User userInMemory=SessionManager.getSession(token);
+        if(userInMemory!=null){
+        	dataWrapper=messageFileDao.getMessageFileList();
+        	dataWrapper.setErrorCode(ErrorCodeEnum.Error);
+        }else{
+        	dataWrapper.setErrorCode(ErrorCodeEnum.User_Not_Logined);
+        }
+        return dataWrapper;
     }
 
 	@Override
